refactor: share the document engine between editor and cloud worker
